Fix false warning about setters being non-public
The accessible flag specifies whether or not the usual java access
checks are going to be ignored for this method, not whether the
method is accessible to the caller.

This flag starts out as false, so all setters were being flagged as
incorrect. Instead, use the modifiers to check whether or not the
setter is actually public.